organize fixed files
